MAINT: Organise the capillary and the surface
