+++ a_package/physics/surfaces.py
@@ -4,12 +4,15 @@ Surface geometry generators.
 Each generator creates a height field on a given grid based on surface configuration.
 """
 
-from dataclasses import asdict
+import dataclasses as dc
 
 import numpy as np
+import numpy.linalg as la
+import numpy.fft as fft
 import numpy.random as random
 
 from a_package.grid import Grid
+from a_package.field import Field, field_component_ax
 from a_package.config.schema import (
     SurfaceConfig,
     FlatSurface,
@@ -18,7 +21,53 @@ from a_package.config.schema import (
     RoughSurface,
     PatternSurface,
 )
-from a_package.physics.models import SelfAffineRoughness, psd_to_height
+
+
+@dc.dataclass(init=True)
+class SelfAffineRoughness:
+    C0: float
+    """Prefactor"""
+    qR: float
+    """Roll-off (angular) wavenumber"""
+    qS: float
+    """Cut-off (angular) wavenumber"""
+    H: float
+    """Hurst exponent"""
+
+    def mapto_isotropic_psd(self, q: Field):
+        """
+        Get the isotropic power spectral density (psd) of a given wavenumber
+        - q: wavenumber in radius, i.e. 2 pi over wavelength
+        """
+        # isotropic, only the magnitude matters
+        wavenumber = la.norm(q, ord=2, axis=field_component_ax, keepdims=True)
+
+        # Find three regimes
+        constant = wavenumber < self.qR
+        self_affine = (wavenumber >= self.qR) & (wavenumber < self.qS)
+        omitted = wavenumber >= self.qS
+
+        # Evaluate accordingly
+        psd = np.full_like(wavenumber, np.nan)
+        psd[constant] = self.C0 * self.qR ** (-2 - 2 * self.H)
+        psd[self_affine] = self.C0 * wavenumber[self_affine] ** (-2 - 2 * self.H)
+        psd[omitted] = 0
+
+        # Return both in convenience of plotting
+        return wavenumber, psd
+
+
+def psd_to_height(psd: Field, rng=None, seed=None):
+    # <h^2> corresponding to <PSD>, thus, take the square-root to match overall amplitude
+    h_amp = np.sqrt(psd)
+
+    # impose some random phase angle following uniform distribution
+    if rng is None:
+        rng = random.default_rng(seed)
+    phase_angle = np.exp(1j * rng.uniform(0, 2 * np.pi, psd.shape))
+
+    # only the sinusoidal is needed
+    return fft.ifft2(h_amp * phase_angle).real
 
 
 def generate_surface(grid: Grid, surface: SurfaceConfig) -> np.ndarray:
